+++ server/views/bookmarklet.js
@@ -7,7 +7,8 @@ function titom1(){
   try{
     if(!b)throw(0);
     d.title='(Tossing...) '+d.title;
-    s.setAttribute('src',l.protocol+'//{{ host }}/toss?u='+encodeURIComponent(l.href)+'&i='+encodeURIComponent(t)+'&t={{ token }}');
+    s.setAttribute('id','titom1');
+    s.setAttribute('src',l.protocol+'//{{ host }}/toss?s=titom1&u='+encodeURIComponent(l.href)+'&i='+encodeURIComponent(t)+'&t={{ token }}');
     b.appendChild(s);
   }catch(e){
     alert('Please wait until the page has loaded.');