+++ CHANGELOG.md
@@ -6,6 +6,7 @@ v2.0.0
 ### Server (2.0.0)
 * Updates to the styling
 * Accepts POSTs from extension and bookmarklet
+* Removes `script` added on bookmarklet initialization
 
 ### Extension (2.0.0)
 * Requests login via POST